+++ back/src/main/java/com/back/domain/member/member/service/MemberStorage.java
@@ -1,0 +1,33 @@
+package com.back.domain.member.member.service;
+
+import com.back.domain.member.member.entity.Member;
+import com.back.domain.member.member.error.MemberErrorCode;
+import com.back.domain.member.mentee.entity.Mentee;
+import com.back.domain.member.mentee.repository.MenteeRepository;
+import com.back.domain.member.mentor.entity.Mentor;
+import com.back.domain.member.mentor.repository.MentorRepository;
+import com.back.global.exception.ServiceException;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public class MemberStorage {
+
+    private final MentorRepository mentorRepository;
+    private final MenteeRepository menteeRepository;
+
+    public Mentor findMentorByMember(Member member) {
+        return findMentorByMemberId(member.getId());
+    }
+
+    public Mentor findMentorByMemberId(Long memberId) {
+        return mentorRepository.findByMemberId(memberId)
+            .orElseThrow(() -> new ServiceException(MemberErrorCode.NOT_FOUND_MENTOR));
+    }
+
+    public Mentee findMenteeByMember(Member member) {
+        return menteeRepository.findByMemberId(member.getId())
+            .orElseThrow(() -> new ServiceException(MemberErrorCode.NOT_FOUND_MENTEE));
+    }
+}
