+++ back/src/main/java/com/back/domain/mentoring/mentoring/service/MentoringStorage.java
@@ -1,10 +1,5 @@
 package com.back.domain.mentoring.mentoring.service;
 
-import com.back.domain.member.member.entity.Member;
-import com.back.domain.member.member.error.MemberErrorCode;
-import com.back.domain.member.mentee.entity.Mentee;
-import com.back.domain.member.mentee.repository.MenteeRepository;
-import com.back.domain.member.mentor.repository.MentorRepository;
 import com.back.domain.mentoring.mentoring.entity.Mentoring;
 import com.back.domain.mentoring.mentoring.error.MentoringErrorCode;
 import com.back.domain.mentoring.mentoring.repository.MentoringRepository;
@@ -20,17 +15,9 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class MentoringStorage {
 
-    private final MentorRepository mentorRepository;
-    private final MenteeRepository menteeRepository;
-
     private final MentoringRepository mentoringRepository;
     private final MentorSlotRepository mentorSlotRepository;
     private final ReservationRepository reservationRepository;
-
-    public Mentee findMenteeByMember(Member member) {
-        return menteeRepository.findByMemberId(member.getId())
-            .orElseThrow(() -> new ServiceException(MemberErrorCode.NOT_FOUND_MENTEE));
-    }
 
     public Mentoring findMentoring(Long mentoringId) {
         return mentoringRepository.findById(mentoringId)
